+++ src/Constants/ChatActions.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace TelegramBot\Constants;
+
+class ChatActions
+{
+    const TYPYNG = 'typing';
+    const UPLOAD_PHOTO = 'upload_photo';
+    const RECORD_VIDEO = 'record_video';
+    const UPLOAD_VIDEO = 'upload_video';
+    const RECORD_AUDIO = 'record_audio';
+    const UPLOAD_AUDIO = 'upload_audio';
+    const UPLOAD_DOCUMENT = 'upload_document';
+    const FIND_LOCATION = 'find_location';
+    const RECORD_VIDEO_NOTE = 'record_video_note';
+    const UPLOAD_VIDEO_NOTE = 'upload_video_note';
+}