+++ src/Constants/ParseModes.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TelegramBot\Constants;
+
+class ParseModes
+{
+    const MARKDOWN = 'markdown';
+    const HTML = 'html';
+}