+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.5.0]
+### Added
+- Added new `endpoint` method to call api methods manually
+- Added 2 classes in constants namespace:
+    - *ChatActions* to use (if you want) in `action` parameter (sendChatAction method)
+    - *ParseModes* to use (if you want) in `parse_mode` parameter
+- Auto split very long text in `sendMessage` method (4096 characters per message)
+    - Enable it after class instantiation:
+      ```php
+      $bot = new TelegramBot($token);
+      $bot->splitLongMessage=true;
+      ```
+    - Default: `false`
+    - Return `Message[]`
+    
 ## [1.4.4]
 ### Added
 - Added license file